+++ src/main/java/com/maan/eway/bean/MsVehicleDetails.java
@@ -339,10 +339,10 @@ private static final long serialVersionUID = 1L;
 	private String previousInsuranceYN;
 	
 	@Column(name = "HORSE_POWER")
-	private Double horsePower;
+	private Integer horsePower;
 	
 	@Column(name = "ZONE")
-	private Double zone;
+	private Integer zone;
 }
